# w2m-ms-spaceships-api-api-specs/src/main/resources/api-contract.yaml
openapi: 3.0.2
info:
  title: Api de naves espaciales
  description: Operaciones para gestionar y buscar naves espaciales de películas y series
  version: 1.0.0
servers:
  - url: /api/v3
tags:
  - name: ApiSpaceship
    description: Operaciones para gestionar naves espaciales

paths:
  /spaceships:
    get:
      tags:
        - ApiSpaceship
      summary: Obtener listado paginado de naves espaciales, admite filtros
      description: Obtiene un listado paginado de todas las naves espaciales, se pueden aplicar filtros
      operationId: searchSpaceships
      security:
        - BearerAuth: []
      parameters:
        - name: name
          in: query
          description: Nombre completo o parcial de la nave
          required: false
          schema:
            type: string
            example: Wing
        - name: page
          in: query
          description: número de paginas a obtener
          required: false
          schema:
            type: number
            default: 0
            minimum: 0
            example: 0
        - name: size
          in: query
          description: número de elementos por página
          required: false
          schema:
            type: number
            default: 10
            minimum: 1
            example: 10
      responses:
        200:
          description: Obtención de datos satisfactoria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpaceshipPage'
        '404':
          description: Nave espacial no encontrada

    post:
      tags:
        - ApiSpaceship
      summary: Crear nave
      description: Crea una nueva nave espacial generando un nuevo Id siempre.
      operationId: createSpaceship
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSpaceship'
      responses:
        200:
          description: Nave espacial creada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Spaceship'

    patch:
      tags:
        - ApiSpaceship
      summary: Modifica nave espacial por Id
      description: Modifica total o parcialmente una nave espacial
      operationId: updateSpaceship
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Spaceship'
      responses:
        '200':
          description: Nave espacial modificada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Spaceship'

  /spaceships/{spaceshipId}:
    get:
      tags:
        - ApiSpaceship
      summary: Obtener nave por Id
      description: Obtiene los detalles de una nave espacial usando su identificador
      operationId: getSpaceshipById
      security:
        - BearerAuth: []
      parameters:
        - name: spaceshipId
          in: path
          required: true
          description: Identificador único de la nave espacial
          schema:
            type: number
            example: 1
      responses:
        '200':
          description: Nave espacial encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Spaceship'
        '404':
          description: Nave espacial no encontrada
    delete:
      tags:
        - ApiSpaceship
      summary: Elimina nave espacial por Id
      description: Elimina definitivamente una nave espacial
      operationId: deleteSpaceship
      security:
        - BearerAuth: []
      parameters:
        - name: spaceshipId
          in: path
          required: true
          description: Identificador único de la nave espacial
          schema:
            type: number
            example: 1
      responses:
        '200':
          description: Nave espacial eliminada correctamente
        '404':
          description: Nave espacial no encontrada

security:
  - BearerAuth: []

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Utiliza el token JWT para autenticar las solicitudes.
  schemas:
    PageInfo:
      properties:
        pageNumber:
          type: number
          description: número de página actual
          example: 1
        pageSize:
          type: number
          description: número de elementos por página
          example: 10
        totalPages:
          type: number
          description: número total de páginas
          example: 2
        totalSize:
          type: number
          description: número total de elementos
          example: 20
    CreateSpaceship:
      properties:
        name:
          type: string
          description: nombre de la nave.
          example: X-Wing Starfighter
    Spaceship:
      properties:
        id:
          type: number
          description: identificador nave.
          example: 1
        name:
          type: string
          description: nombre de la nave.
          example: X-Wing Starfighter
    SpaceshipPage:
      properties:
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
        spaceships:
          type: array
          items:
            $ref: '#/components/schemas/Spaceship'